moved get admin and get users to db file
